Move some items from storage into storage::mvcc/txn
