+++ src/storage/mvcc/lock.rs
@@ -1,9 +1,11 @@
 // Copyright 2016 TiKV Project Authors. Licensed under Apache-2.0.
 
 use crate::storage::{
-    mvcc::{ErrorInner, Result, TimeStamp, TsSet},
-    Mutation, FOR_UPDATE_TS_PREFIX, MIN_COMMIT_TS_PREFIX, SHORT_VALUE_MAX_LEN, SHORT_VALUE_PREFIX,
-    TXN_SIZE_PREFIX,
+    mvcc::{
+        ErrorInner, Result, TimeStamp, TsSet, FOR_UPDATE_TS_PREFIX, MIN_COMMIT_TS_PREFIX,
+        SHORT_VALUE_PREFIX, TXN_SIZE_PREFIX,
+    },
+    Mutation, SHORT_VALUE_MAX_LEN,
 };
 use byteorder::ReadBytesExt;
 use derive_new::new;
