+++ src/storage/mvcc/write.rs
@@ -1,8 +1,7 @@
 // Copyright 2016 TiKV Project Authors. Licensed under Apache-2.0.
 
 use super::lock::LockType;
-use super::{Error, ErrorInner, Result, TimeStamp};
-use crate::storage::{SHORT_VALUE_MAX_LEN, SHORT_VALUE_PREFIX};
+use super::{Error, ErrorInner, Result, TimeStamp, SHORT_VALUE_MAX_LEN, SHORT_VALUE_PREFIX};
 use codec::prelude::NumberDecoder;
 use keys::Value;
 use tikv_util::codec::number::{NumberEncoder, MAX_VAR_U64_LEN};
