+++ src/storage/txn/commands.rs
@@ -7,15 +7,14 @@ use kvproto::kvrpcpb::{CommandPri, Context, GetRequest, RawGetRequest};
 use tikv_util::collections::HashMap;
 
 use crate::storage::metrics::{self, CommandPriority};
-use crate::storage::mvcc::{Lock, TimeStamp};
-use crate::storage::Mutation;
+use crate::storage::mvcc::{Lock, Mutation, TimeStamp};
 
 /// Get a single value.
 pub struct PointGetCommand {
-    pub(super) ctx: Context,
-    pub(super) key: Key,
+    pub ctx: Context,
+    pub key: Key,
     /// None if this is a raw get, Some if this is a transactional get.
-    pub(super) ts: Option<TimeStamp>,
+    pub ts: Option<TimeStamp>,
 }
 
 impl PointGetCommand {
